+++ Scribble.Shared/Lib/Event.cs
@@ -1,7 +1,5 @@
 using System.Text.Json.Serialization;
-using Scribble.Shared.Lib.CanvasElements;
 using Scribble.Shared.Lib.Events;
-using SkiaSharp;
 
 namespace Scribble.Shared.Lib;
 
@@ -53,17 +51,3 @@ public abstract record Event(Guid ActionId)
 public interface ITerminalEvent
 {
 }
-
-// MISC
-public record LoadCanvasEvent(Guid ActionId, List<CanvasElement> CanvasElements) : Event(ActionId);
-
-public record UndoEvent(Guid ActionId, Guid TargetActionId) : Event(ActionId);
-
-public record RedoEvent(Guid ActionId, Guid TargetActionId) : Event(ActionId);
-
-public record PasteCanvasElementsEvent(
-    Guid ActionId,
-    SKPoint Position,
-    List<CanvasElement> CopiedElements,
-    Guid SelectionBoundId)
-    : Event(ActionId), ITerminalEvent;
