+++ Scribble/ViewModels/MainViewModel.cs
@@ -12,6 +12,7 @@ using Scribble.Services.DialogService;
 using Scribble.Services.MultiUserDrawing;
 using Scribble.Shared.Lib;
 using Scribble.Shared.Lib.CanvasElements;
+using Scribble.Shared.Lib.Events;
 using SkiaSharp;
 
 namespace Scribble.ViewModels;
